switched to floating point, added string to assembler

+++ awtea-graphics/src/main/java/me/mdbell/awtea/util/ast/AstArg.java
@@ -9,5 +9,5 @@ import lombok.ToString;
 @ToString
 public class AstArg {
     private final String text;
-    private final Integer value;
+    private final Float value;
 }
